refactor(forks): move State and Store bodies into the spec class
Migrates every State and Store method body into LstarSpec. Containers
(State, Store, SignedBlock) become thin Pydantic data classes whose
methods are one-line forwarders to the active fork spec, reached via a
deferred import that breaks the spec ↔ container module-load cycle.

Inside the moved bodies, every literal Block(...), BlockBody(...),
BlockHeader(...), Config(...), AggregatedAttestations(...), and other
container constructor is now self.<name>_class(...). An inheriting fork
that swaps a single container type therefore receives the parent fork's
logic for free.

Observability hooks (observe_state_transition, observe_on_block,
observe_on_attestation) ride along with the bodies, preserving the
metrics surface.

The obsolete delegator-forwarding test file is removed; behavioural
coverage now lives in the existing state-transition, fork-choice, and
block-production test suites.
